requirements.txt
aiohttp==3.8.4
aiosqlite==0.19.0
discord.py==2.3.0
nest_asyncio==1.5.6
openai==0.27.8
yt_dlp==2023.6.22
